Dependabot/requirements.txt
numpy==1.16.0
requests==2.20.0
PyYAML==6.0.3
#libary=="Version) Oben genante konfigurations sind auf den Testcode zugeschnitten, können geändert werden
